requirements.txt
flask==2.3.3
flask-babel==4.0.0
twilio==8.2.0
openai==0.28.1
deep-translator==1.11.4
requests==2.31.0
python-dotenv==1.0.0
# sqlite3
schedule==1.2.0
flask-cors==4.0.0
pandas==2.0.3
nltk==3.8.1
scikit-learn==1.3.0
langdetect==1.0.9
APScheduler==3.10.1
python-telegram-bot==20.4
flask-sqlalchemy==3.0.5
cryptography==41.0.3
jsonify
datetime